Add parameter covariance diagnostics, fix division by zero.,
