+++ src/CalibrateEDMF.jl
@@ -7,6 +7,7 @@ include("ReferenceModels.jl")
 include("LESUtils.jl")
 include("ReferenceStats.jl")
 include("TurbulenceConvectionUtils.jl")
+include("Diagnostics.jl")
 include("NetCDFIO.jl")
 include("Pipeline.jl")
